# Fix parsing issues in models.scanKey