@e-dard 2015-11-19
I also plan on fixing the lint issues in the `models` package, but will wait until the current changes have been reviewed.