@jwilder 2015-11-19
This parsing code has been pretty highly tuned because it's used in a lot of places (HTTP/UDP inputs, hinted handoff, WAL, cluster writes, exports, etc..).  The performance decrease is fairly significant because it reduces throughput by 20-45% in common cases.

```
name                         old time/op    new time/op    delta
Marshal-4                      6.11µs ± 0%    6.14µs ± 0%   +0.62%
ParsePointNoTags-4              844ns ± 0%     838ns ± 0%   -0.71%
ParsePointsTagsSorted2-4       1.45µs ± 0%    1.48µs ± 0%   +2.00%
ParsePointsTagsSorted5-4       1.70µs ± 0%    2.07µs ± 0%  +21.76%
ParsePointsTagsSorted10-4      2.29µs ± 0%    3.32µs ± 0%  +44.72%
ParsePointsTagsUnSorted2-4     1.47µs ± 0%    1.76µs ± 0%  +20.26%
ParsePointsTagsUnSorted5-4     1.55µs ± 0%    2.17µs ± 0%  +39.54%
ParsePointsTagsUnSorted10-4    3.93µs ± 0%    4.99µs ± 0%  +26.91%

name                         old speed      new speed      delta
ParsePointNoTags-4           27.2MB/s ± 0%  27.4MB/s ± 0%   +0.62%
ParsePointsTagsSorted2-4     35.1MB/s ± 0%  34.4MB/s ± 0%   -1.91%
ParsePointsTagsSorted5-4     48.8MB/s ± 0%  40.1MB/s ± 0%  -17.85%
ParsePointsTagsSorted10-4    62.4MB/s ± 0%  43.1MB/s ± 0%  -30.89%
ParsePointsTagsUnSorted2-4   34.8MB/s ± 0%  28.9MB/s ± 0%  -16.90%
ParsePointsTagsUnSorted5-4   53.4MB/s ± 0%  38.3MB/s ± 0%  -28.34%
ParsePointsTagsUnSorted10-4  36.4MB/s ± 0%  28.7MB/s ± 0%  -21.18%
```

The `UnSorted` cases would affect ingestion endpoints that parse line protocol such as HTTP/UDP.  The `Sorted` cases affect all the other parts of the system that are using the the line protocol as a serialization format after already being parsed by an input plugin.

I have not profiled this change, but I suspect the performance decrease is due to that closures and calling a func per char whereas before it would just increment an integer.  The code doesn't seem much clearer to me either unfortunately, but I see where you were going with it.  I'd suggest reworking this to have `scanKey` internally call a `scanMeasurement` and `scanTags`.  `scanTags` could call a `scanTagKey` and `scanTagValue`, etc. if needed.  Each of those would have slice passed in similar to how the funcs in the package work.  That is the pattern used in the other funcs (at times).

I'm :-1: on this current PR, but I think it could be reworked to address these concerns and fix the original parser bugs.