@e-dard 2015-11-19
Hi @jwilder, your points seem completely reasonable to me.

This is my first contribution to the project and as you can imagine I have a shallow understanding of the entire codebase, and the impact of `scanKeys` on it, so your feedback is useful.

I will profile and see what I can do to get the benchmarks up to their previous levels :+1: 

If that's unsuccessful, then I have another [branch waiting](https://github.com/e-dard/influxdb/commit/27e46376fbac07052c9527ce0ed193660e06cbff) that will simply fix the outstanding issues.