@e-dard 2015-11-20
@jwilder,

I've modified the approach to bring performance inline with expectation, while keeping separation of the different parts of the scanning process in different functions.

And of course the bugs should still be fixed :smile: 

Results of six runs of the following benchmark command: `$ go test -benchtime=3s -bench .` for `master` and this PR.

```
name                         old time/op    new time/op    delta
Marshal-4                      6.08µs ± 1%    6.10µs ± 1%     ~     (p=0.132 n=6+6)
ParsePointNoTags-4              838ns ± 0%     828ns ± 1%   -1.23%  (p=0.004 n=5+6)
ParsePointsTagsSorted2-4       1.16µs ± 2%    1.14µs ± 0%   -1.95%  (p=0.010 n=6+4)
ParsePointsTagsSorted5-4       1.52µs ± 1%    1.42µs ± 0%   -6.22%  (p=0.004 n=6+5)
ParsePointsTagsSorted10-4      2.28µs ± 1%    2.07µs ± 0%   -9.45%  (p=0.004 n=6+5)
ParsePointsTagsUnSorted2-4     1.42µs ± 0%    1.41µs ± 1%   -1.18%  (p=0.010 n=4+6)
ParsePointsTagsUnSorted5-4     1.55µs ± 1%    1.44µs ± 0%   -7.24%  (p=0.002 n=6+6)
ParsePointsTagsUnSorted10-4    3.87µs ± 1%    3.59µs ± 1%   -7.22%  (p=0.002 n=6+6)

name                         old speed      new speed      delta
ParsePointNoTags-4           27.4MB/s ± 0%  27.8MB/s ± 1%   +1.24%  (p=0.004 n=5+6)
ParsePointsTagsSorted2-4     43.8MB/s ± 2%  44.6MB/s ± 0%   +1.98%  (p=0.010 n=6+4)
ParsePointsTagsSorted5-4     54.7MB/s ± 1%  58.3MB/s ± 0%   +6.62%  (p=0.004 n=6+5)
ParsePointsTagsSorted10-4    62.6MB/s ± 1%  69.1MB/s ± 0%  +10.44%  (p=0.004 n=6+5)
ParsePointsTagsUnSorted2-4   35.8MB/s ± 0%  36.2MB/s ± 1%   +1.29%  (p=0.004 n=5+6)
ParsePointsTagsUnSorted5-4   53.6MB/s ± 1%  57.8MB/s ± 0%   +7.79%  (p=0.002 n=6+6)
ParsePointsTagsUnSorted10-4  36.9MB/s ± 1%  39.8MB/s ± 1%   +7.77%  (p=0.002 n=6+6)
```

So performance is now on a par or slightly better than before.

Thoughts?