@jwilder 2015-11-20
@e-dard Nice work! Not only fixed the bugs, but code is simpler and faster.

One question, but this looks great. :+1: