@e-dard 2015-11-20
@jwilder couple of things:

 1. I will add a commit that lints the package.
 2. I've found other places in the file where there are potentially escape sequence bugs, along the same lines as the ones I've fixed. Obviously the suite is passing so it's not clear if they can get triggered. Example below:

Consider `scanTo` for example, if it's given an input of a literal backslash followed by an escaped stop character it will not return the expected token (at least as I understand the purpose of `scanTo`).

[Playground](http://play.golang.org/p/R-zdOwhxt6)

I think the file could do with a review of all the functions that are incorrectly skipping `buf` when they see a literal slash. Maybe I should do this in another issue/PR? What do you think?