@e-dard 2015-11-20
@otoolep I did when I contributed to `influxdb/influxdb.com`. Do I need to sign a different one or is it the same CLA?