@otoolep 2015-11-20
As long as you signed https://influxdb.com/community/cla.html we're good.